Add ShoppingCart, service with their interfaces, fix Price type in MotorOil interface, add EF package to Domain.Interfaces

+++ src/eAuto.Domain.Interfaces/IMotorOil.cs
@@ -5,7 +5,7 @@
         int MotorOilId { get; set; }
         string Name { get; set; }
         string PictureUrl { get; set; }
-        decimal Price { get; set; }
+        double Price { get; set; }
         string Viscosity { get; set; }
 		string Composition { get; set; }
 		int Volume { get; set; }
